chore: rename user entities to repository
