+++ apps/backend/src/user/user.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserEntitiesModule } from './entities/user-entities.module';
+import { UserEntitiesModule } from './repositories/user-repository.module';
 import { UserService } from './user.service';
 
 @Module({
